flaskauth/requirements.txt
Flask~=3.0.2
Flask-SQLAlchemy
wheel
pyjwt~=2.8.0
datetime~=5.4
uuid
pytest==7.1.2
coverage==6.4.2
python-dotenv~=1.0.1
alembic~=1.13.1
mysqlclient
flask-marshmallow
marshmallow-sqlalchemy
cerberus~=1.3.5
Flask-Mail
celery~=5.3.6
redis
requests
SQLAlchemy~=2.0.25
setuptools~=69.0.3
Werkzeug~=3.0.1
jsonschema~=4.21.1